requirements.txt
pinecone-client==2.2.2
transformers==4.31.0
torch==2.0.1
PyPDF2==3.0.1
fastapi==0.101.0
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
